revert(core): the meta's entry snapshot cost ≈15% on the hottest arc (#1722)
- 0.89.9 read the meta's three flags off a snapshot taken at the entry rather
  than off the caller's opts inside the commit; navigate/sync-baseline went
  8.3 → 9.8ms and navigate/pre-commit-listener 5.8 → 6.5, everything else
  unchanged
- six runner configurations rule out the plan literal's width, the call shape,
  the entry reads, field count as such and the runner itself; the only line that
  splits them is where the flags are read from, and the mechanism is not
  understood
- the ordering rule and commit-ask-snapshot-1649 come back with it
- #1717 is untouched: the commit gate asking the captured signal measured clean
