+++ ARCHITECTURE.md
@@ -321,7 +321,7 @@ FSM events trigger observable emissions through two paths:
 
 Cost: three transition-payload allocations per navigation (`NAVIGATE`, `LEAVE_APPROVE`, `COMPLETE` each carry one) — a deliberate trade of micro-optimization for structural determinism. Each payload carries only what something downstream reads: `CANCEL` and `FAIL` name no target, because their actions take it from the context. Correctness is enforced by the state machine rather than by scattered re-checks — `when: mayCommit` on the `COMPLETE` edge is the whole commit gate.
 
-**The commit is asked of the table ONCE, and the position is load-bearing.** `completeTransition` runs one piece of destructive bookkeeping on its way to the commit — the post-leave cleanup that unregisters the departing route's EXTERNAL `canDeactivate` — and the ask stands **above** it. That ordering is the whole guarantee: a navigation the table refuses stops before the cleanup, so it cannot unregister the guard of the route the user is staying on. Nothing else runs between the verdict and the send. The cleanup is pure bookkeeping — clearing a guard re-derives the compiled slot by READING the surviving origin's stored compiled form, never by invoking its factory — so no application code exists in that window to invalidate a verdict already given. Nor does the function reach the caller's options object at all: the three flags the `TransitionMeta` is built from are read once at the navigation's entry and travel on its plan, so the emptiness is structural and not a matter of where the meta is built. That is what makes a single snapshot verdict sound, because the send cannot report back — its action announces the success synchronously, so a listener's own commit is indistinguishable from a refusal.
+**The commit is asked of the table ONCE, and the position is load-bearing.** `completeTransition` runs one piece of destructive bookkeeping on its way to the commit — the post-leave cleanup that unregisters the departing route's EXTERNAL `canDeactivate` — and the ask stands **above** it. That ordering is the whole guarantee: a navigation the table refuses stops before the cleanup, so it cannot unregister the guard of the route the user is staying on. Nothing else runs between the verdict and the send. The cleanup is pure bookkeeping — clearing a guard re-derives the compiled slot by READING the surviving origin's stored compiled form, never by invoking its factory — so no application code exists in that window to invalidate a verdict already given. The one step that DOES touch application code, building the `TransitionMeta` from the caller's own options object, sits above the ask for exactly that reason: a single snapshot verdict is sound only while everything below it is inert, and the send cannot report back — its action announces the success synchronously, so a listener's own commit is indistinguishable from a refusal.
 
 **The table also OWNS the committed state.** `current` / `previous` are fields of the FSM context written by the four edge `update`s above and by nothing else, so "committed" and "announced" cannot come apart — the silent commit (state written, `TRANSITION_SUCCESS` never emitted, no subscriber notified) stops being expressible rather than being guarded against (`packages/core/INVARIANTS.md`, "Committed-state ownership"). One observable consequence: the pair is zeroed on the `DISPOSE` edge, which runs before plugin teardown, so a plugin `teardown()` reading `getPreviousState()` sees `undefined`.
 
@@ -467,7 +467,7 @@ Guard order is fixed: deactivation innermost → outermost, then activation oute
 
 1. Re-check that the target route still exists — route CRUD can have removed it mid-flight — otherwise `ROUTE_NOT_FOUND` through `FAIL`.
 2. Put the commit to the table with the navigation's own context AS the payload — no second literal — carrying `opts` **unstripped**, because the announcement hands them to every plugin and sanitising them for subscribers is that action's job. The table never reads them: the commit predicate and the action's strip branch both ask the external signal the navigation captured at its entry, which is a field of the same payload.
-3. Attach and freeze the `TransitionMeta`, freeze the state. Its three flags come off the navigation's PLAN, where the entry point put them before anything was announced — so this step, like the rest of the function, runs no application code and needs no position relative to the verdict.
+3. Attach and freeze the `TransitionMeta`, freeze the state. This step reads three flags off the CALLER's options object, which may be accessor- or Proxy-backed, so it is the last application code in the function and it is deliberately above the verdict.
 4. **Ask the table**, once and unconditionally. Everything that can move the router has already run, and everything after this point is inert — that is what makes one snapshot verdict sound.
 5. Run the post-leave cleanup, which unregisters the departing route's external `canDeactivate`. It is destructive, so it is below the ask: a navigation the table refuses must not take the guard of the route the user is staying on with it. `clearCanDeactivate` demands the permit the ask returned, so the two cannot be reordered.
 6. **Fire** — `send(COMPLETE)`'s `update` writes `current` / `previous`; its action emits `TRANSITION_SUCCESS`.
